server/requirements.txt
Flask==1.1.2
redis==3.4.1
gunicorn>=20
Flask-RESTful==0.3.8
Flask-Cors==3.0.8
werkzeug<=1.0.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
Flask-Script<=2.0.6
numpy==1.14.3
Pillow<=4.1.1
imageio-ffmpeg
moviepy<=0.2.2.11
Flask-Passlib==0.1
Flask-HTTPAuth==4.0.0
passlib<=1.7.2
itsdangerous==1.1.0
